crud course

+++ backend/src/main/java/com/fsa/cursus/model/response/CategoryResponse.java
@@ -8,7 +8,7 @@ import java.io.Serializable;
 @Getter
 @Setter
 public class CategoryResponse implements Serializable {
-    private int categoryId;
+    private Long categoryId;
     private String categoryName;
     private Boolean categoryStatus;
 }
